# public/docs/dash/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Documentation for the REST API of Petnet Dashboard.",
    "version": "1.0.0",
    "title": "Petnet Dashboard API",
    "termsOfService": "",
    "contact": {
      "email": "akrambelhajsalem@cyberdivision.xyz"
    }
  },
  "host": "dev-api.pet-net.com.au",
  "basePath": "/api/v1",
...
  "schemes": ["https"],
  "paths": {
...
    "/admin/user": {
      "get": {
        "tags": ["user"],
        "summary": "Get Current Admin",
        "description": "Get authenticated admin profile.",
        "operationId": "getCurrentAdmin",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetCurrentAdminRes"
            }
          },
          "500": {
            "description": "Something Went Wrong",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/admin/users-all": {
      "get": {
        "tags": ["users"],
        "summary": "Get Users",
        "description": "Get Users in chat .",
        "operationId": "getUsers",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetUsersSearch"
            }
          },
          "500": {
            "description": "Something Went Wrong",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/admin/user-update": {
      "put": {
        "tags": ["user"],
        "summary": "Update Current User",
        "description": "Update authenticated user profile.",
        "operationId": "updateCurrentUser",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateCurrentUserReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UpdateCurrentUserRes"
            }
          },
          "400": {
            "description": "Invalid Request",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Something Went Wrong",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/admin/user/password": {
      "post": {
        "tags": ["user"],
        "summary": "Update Password",
        "description": "Update password.",
        "operationId": "updatePassword",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdatePasswordReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UpdatePasswordRes"
            }
          },
          "500": {
            "description": "Something Went Wrong",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/admin/users": {
      "post": {
        "tags": ["users"],
        "summary": "Create User",
        "description": "Create a new user",
        "operationId": "createUser",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateUserReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CreateUserRes"
            }
          },
          "500": {
            "description": "Something Went Wrong",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "get": {
        "tags": ["users"],
        "summary": "Get All Users",
        "description": "Get All Users",
        "operationId": "getAllUsers",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "query",
            "name": "pageNumber",
            "description": "Page Number",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "pageSize",
            "description": "Page Size(items per page)",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "userKind",
            "description": "User Kind Query",
            "required": false,
            "type": "string",
            "enum": ["STANDARD", "BREEDER", "ADMIN"]
          },
          {
            "in": "query",
            "name": "search",
            "description": "Search Post by multiple fields",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetUsersRes"
            }
          },
          "500": {
            "description": "Something Went Wrong",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/admin/users/{userId}": {
      "get": {
        "tags": ["users"],
        "summary": "Get User Profile",
        "description": "Replace the {userId} by the id of the user desired to get profile , all the fields are optional ,just specify the field you want to change ",
        "operationId": "getUserProfile",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "description": "Id of the User",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "pageNumber",
            "description": "Page Number",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "pageSize",
            "description": "Page Size(items per page)",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetUserProfileRes"
            }
          },
          "500": {
            "description": "Something Went Wrong",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "tags": ["users"],
        "summary": "Update User",
        "description": "Replace the {userId} by the id of the user desired to update , all the fields are optional ,just specify the field you want to change ",
        "operationId": "updateUser",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "description": "Id of the User",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateUserReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UpdateUserRes"
            }
          },
          "500": {
            "description": "Something Went Wrong",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "patch": {
        "tags": ["users"],
        "summary": "Update User Activation Status",
        "description": "Replace the {userId} by the id of the user desired to update activation status , all the fields are optional ,just specify the field you want to change ",
        "operationId": "updateUserActivationStatus",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "description": "Id of the User",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateUserActivationStatusReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/UpdateUserActivationStatusRes"
            }
          },
          "500": {
            "description": "Something Went Wrong",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": ["users"],
        "summary": "Delete User",
        "description": "Replace the {userId} by the id of the user desired to delete , all the fields are optional ,just specify the field you want to change ",
        "operationId": "deleteUser",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "description": "Id of the User",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DeleteUserRes"
            }
          },
          "500": {
            "description": "Something Went Wrong",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/admin/users-state": {
      "get": {
        "tags": ["users"],
        "summary": "count users by state",
        "description": "count users by state",
        "operationId": "countUsersByState",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CountUsersByStateRes"
            }
          },
          "500": {
            "description": "Something Went Wrong",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "Error": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "UpdatePasswordReq": {
      "type": "object",
      "required": ["oldPassword", "newPassword"],
      "properties": {
        "oldPassword": {
          "type": "string"
        },
        "newPassword": {
          "type": "string"
        }
      }
    },
    "UpdatePasswordRes": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
...
    "GetCurrentAdminRes": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "account": {
          "type": "object",
          "properties": {
            "kind": {
              "type": "string"
            },
            "isVerified": {
              "type": "boolean"
            },
            "email": {
              "type": "string"
            }
          }
        },
        "profile": {
          "type": "object",
          "properties": {
            "firstName": {
              "type": "string"
            },
            "lastName": {
              "type": "string"
            },
            "phone": {
              "type": "string"
            },
            "state": {
              "type": "string"
            },
            "suburb": {
              "type": "string"
            },
            "address": {
              "type": "string"
            },
            "zipCode": {
              "type": "integer"
            },
            "avatarUrl": {
              "type": "string"
            }
          }
        }
      }
    },
...
    "GetUsersRes": {
      "type": "object",
      "properties": {
        "total": {
          "type": "integer",
          "format": "int32"
        },
        "users": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "userId": {
                "type": "string"
              },
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              },
              "avatarUrl": {
                "type": "string"
              },
              "kind": {
                "type": "string"
              },
              "activationStatus": {
                "type": "string",
                "enum": ["ACTIVE", "INACTIVE", "SUSPENDED", "BLOCKED"]
              },
              "isVerified": {
                "type": "boolean"
              },
              "registrationMethod": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "phone": {
                "type": "string"
              }
            }
          }
        }
      }
    },
...
    "CountUsersByStateRes": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "output": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string"
              },
              "count": {
                "type": "integer"
              }
            },
            "required": ["_id", "count"]
          }
        }
      }
    },
...
    "CreateUserReq": {
      "type": "object",
      "required": ["kind", "isVerified", "email", "firstName", "lastName"],
      "properties": {
        "kind": {
          "type": "string",
          "enum": ["STANDARD", "BREEDER", "ADMIN"]
        },
        "isVerified": {
          "type": "boolean"
        },
        "email": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "suburb": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "zipCode": {
          "type": "integer",
          "format": "int32"
        },
        "avatarUrl": {
          "type": "string"
        }
      }
    },
    "CreateUserRes": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "userId": {
          "type": "string"
        }
      }
    },
    "DeleteUserRes": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
    "UpdateCurrentUserReq": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "suburb": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "zipCode": {
          "type": "integer",
          "format": "int32"
        },
        "bio": {
          "type": "string"
        },
        "avatarUrl": {
          "type": "string"
        },
        "deviceId": {
          "type": "string"
        }
      }
    },
    "UpdateCurrentUserRes": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
    "UpdateUserActivationStatusReq": {
      "type": "object",
      "properties": {
        "activationStatus": {
          "type": "string",
          "enum": ["ACTIVE", "INACTIVE", "SUSPENDED", "BLOCKED"]
        }
      }
    },
    "UpdateUserActivationStatusRes": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
    "GetUserProfileRes": {
      "type": "object",
      "properties": {
        "user": {
          "type": "object",
          "properties": {
            "userId": {
              "type": "string"
            },
            "account": {
              "type": "object",
              "properties": {
                "kind": {
                  "type": "string"
                },
                "activationStatus": {
                  "type": "string",
                  "enum": ["ACTIVE", "INACTIVE", "SUSPENDED", "BLOCKED"]
                },
                "isVerified": {
                  "type": "boolean"
                },
                "registrationMethod": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "lastSignIn": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            },
            "profile": {
              "type": "object",
              "properties": {
                "firstName": {
                  "type": "string"
                },
                "lastName": {
                  "type": "string"
                },
                "phone": {
                  "type": "string"
                },
                "state": {
                  "type": "string"
                },
                "suburb": {
                  "type": "string"
                },
                "address": {
                  "type": "string"
                },
                "zipCode": {
                  "type": "integer",
                  "format": "int32"
                },
                "bio": {
                  "type": "string"
                },
                "avatarUrl": {
                  "type": "string"
                }
              }
            }
          }
        },
        "posts": {
          "type": "object",
          "properties": {
            "total": {
              "type": "integer",
              "format": "int32"
            },
            "items": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "postId": {
                    "type": "string"
                  },
                  "kind": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "adoptionFee": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "breed": {
                    "type": "string"
                  },
                  "dadBreed": {
                    "type": "string"
                  },
                  "momBreed": {
                    "type": "string"
                  },
                  "location": {
                    "type": "object",
                    "properties": {
                      "state": {
                        "type": "string"
                      },
                      "suburb": {
                        "type": "string"
                      }
                    }
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "url": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "childImages": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "url": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "isLiked": {
                    "type": "boolean"
                  },
                  "likesCount": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "createdAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        }
      }
    },
    "UpdateUserReq": {
      "type": "object",
      "properties": {
        "kind": {
          "type": "string",
          "enum": ["STANDARD", "BREEDER", "ADMIN"]
        },
        "activationStatus": {
          "type": "string",
          "enum": ["ACTIVE", "INACTIVE", "SUSPENDED", "BLOCKED"]
        },
        "email": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "suburb": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "zipCode": {
          "type": "integer",
          "format": "int32"
        },
        "avatarUrl": {
          "type": "string"
        }
      }
    },
    "UpdateUserRes": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        }
      }
    },
...
    "GetUsersSearch": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "account": {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string"
              }
            }
          },
          "profile": {
            "type": "object",
            "properties": {
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              },
              "avatarUrl": {
                "type": "string"
              }
            }
          }
        }
      }
    },
...
  }
}